ibrahima changes

+++ app.py
@@ -1,5 +1,5 @@
 import pymysql.cursors
-import dotenv
+
 import os
 from flask import Flask, request, render_template, redirect, url_for, abort, flash
 from flask import session, g
@@ -7,7 +7,7 @@ from flask import session, g
 app = Flask(__name__)
 app.secret_key = 'une cle(token) : grain de sel(any random string)'
 def get_db():
-    dotenv.load_dotenv()
+
     if 'db' not in g:
         g.db = pymysql.connect(
             host="localhost",
@@ -142,7 +142,7 @@ def valid_edit_collecte():
 @app.route('/Tournee/show')
 def show_Tournee():
     mycursor = get_db().cursor()
-    sql = '''SELECT id_tournee, `date-tournee`, id_centre_recyclage, id_camion, temps
+    sql = '''SELECT id_tournee, `date_tournee`, id_centre_recyclage, id_camion, temps
              FROM Tournee'''
     mycursor.execute(sql)
     Tournee = mycursor.fetchall()
@@ -154,21 +154,29 @@ def add_Tournee():
     print('''affichage du formulaire pour saisir une Tournee''')
     return render_template('Tournee/add_Tournee.html')
 
+
 @app.route('/Tournee/delete')
 def delete_Tournee():
-    print('''suppression d'une Tournee''')
-    tournee_id = request.args.get('id', None)
-    print(tournee_id)
-    mycursor = get_db().cursor()
-    tuple_param = (tournee_id,)
-    sql = "DELETE FROM Tournee WHERE id_tournee=%s;"
-    mycursor.execute(sql, tuple_param)
+    print('''Suppression d'une Tournée''')
+    id_tournee = request.args.get('id')
 
-    get_db().commit()
-    print(request.args)
-    print(request.args.get('id'))
-    tournee_id = request.args.get('id', 0)
+    if id_tournee:
+        try:
+            id_tournee = int(id_tournee)
+            mycursor = get_db().cursor()
+            tuple_param = (id_tournee,)
+            sql = "DELETE FROM Tournee WHERE id_tournee=%s;"
+            mycursor.execute(sql, tuple_param)
+            get_db().commit()
+
+            message = f'info: Tournée supprimée - ID : {id_tournee}'
+            flash(message, 'alert-warning')
+        except ValueError:
+            print("L'ID de la tournée n'est pas un entier valide.")
+
     return redirect('/Tournee/show')
+
+
 
 @app.route('/Tournee/edit', methods=['GET'])
 def edit_Tournee():
@@ -185,38 +193,55 @@ def edit_Tournee():
     Tournee = mycursor.fetchone()
     return render_template('Tournee/edit_Tournee.html', tournee=Tournee)
 
+
 @app.route('/Tournee/add', methods=['POST'])
 def valid_add_Tournee():
-    print('''ajout de la tournee dans le tableau''')
+    print('''Ajout de la tournée dans la table''')
     id_tournee = request.form.get('id_tournee')
     date_tournee = request.form.get('date_tournee')
     id_centre_recyclage = request.form.get('id_centre_recyclage')
     id_camion = request.form.get('id_camion')
     temps = request.form.get('temps')
-    message = ' id_tournee :' + id_tournee + ' - date_tournee  :' + date_tournee + ' - id_centre_recyclage  :' + id_centre_recyclage + ' - id_camion : ' + id_camion + ' - temps : ' + temps
-    print(message)
+
+    message = (
+        f'info: Tournée ajoutée - ID : {id_tournee}, Date : {date_tournee}, '
+        f'Centre de recyclage : {id_centre_recyclage}, Camion : {id_camion}, '
+        f'Temps : {temps}'
+    )
+
+    flash(message, 'alert-success')
+
     mycursor = get_db().cursor()
     tuple_param = (id_tournee, date_tournee, id_centre_recyclage, id_camion, temps)
     sql = "INSERT INTO Tournee(id_tournee, date_tournee, id_centre_recyclage, id_camion, temps) VALUES (%s, %s, %s, %s, %s);"
     mycursor.execute(sql, tuple_param)
     get_db().commit()
+
     return redirect('/Tournee/show')
+
 
 @app.route('/Tournee/edit', methods=['POST'])
 def valid_edit_Tournee():
-    print('''modification de la Tournee dans le tableau''')
+    print('''Modification de la Tournée dans le tableau''')
     id_tournee = request.form.get('id_tournee')
     date_tournee = request.form.get('date_tournee')
     id_centre_recyclage = request.form.get('id_centre_recyclage')
     id_camion = request.form.get('id_camion')
     temps = request.form.get('temps')
-    message = ' id_tournee :' + id_tournee + ' - date_tournee  :' + date_tournee + ' - id_centre_recyclage  :' + id_centre_recyclage + ' - id_camion : ' + id_camion + ' - temps : ' + temps
-    print(message)
+
+    message = (
+        f'info: Tournée modifiée - ID : {id_tournee}, Date : {date_tournee}, '
+        f'Centre de recyclage : {id_centre_recyclage}, Camion : {id_camion}, '
+        f'Temps : {temps}'
+    )
+    flash(message, 'alert-success')
+
     mycursor = get_db().cursor()
     tuple_param = (id_tournee, date_tournee, id_centre_recyclage, id_camion, temps, id_tournee)
     sql = "UPDATE Tournee SET id_tournee = %s, date_tournee = %s, id_centre_recyclage = %s, id_camion = %s, temps = %s WHERE id_tournee = %s;"
     mycursor.execute(sql, tuple_param)
     get_db().commit()
+
     return redirect('/Tournee/show')
